Reorganizing STEP library (#302)
re-org of step library for graceful migration to Rules V2
small adjustments in UI
