+++ cognite/neat/workflows/steps/lib/__init__.py
@@ -1,8 +1,7 @@
-from cognite.neat.workflows.steps.lib.graph_extractor import *  # noqa
-from cognite.neat.workflows.steps.lib.graph_transformer import *  # noqa
-from cognite.neat.workflows.steps.lib.graph_loader import *  # noqa
-from cognite.neat.workflows.steps.lib.graph_store import *  # noqa
-from cognite.neat.workflows.steps.lib.rules_exporter import *  # noqa
-from cognite.neat.workflows.steps.lib.graph_contextualization import *  # noqa
-from cognite.neat.workflows.steps.lib.io_steps import *  # noqa
-from cognite.neat.workflows.steps.lib.rules_importer import *  # noqa
+from .io_steps import *  # noqa
+from .rules_importer import *  # noqa
+from .rules_validator import *  # noqa
+from .rules_exporter import *  # noqa
+from .graph_store import *  # noqa
+from .graph_loader import *  # noqa
+from .graph_extractor import *  # noqa
