+++ docs/api/workflows/steps/graph.md
@@ -1,9 +1,9 @@
 # Graph Store Steps
 ::: cognite.neat.workflows.steps.lib.graph_store
-# Graph Extractor Steps
-::: cognite.neat.workflows.steps.lib.graph_extractor
+<!-- # Graph Extractor Steps
+::: cognite.neat.workflows.steps.lib.graph_extractor -->
 # Graph Loader Steps
 ::: cognite.neat.workflows.steps.lib.graph_loader
 
-# Graph Transformer Steps
-::: cognite.neat.workflows.steps.lib.graph_transformer
+<!-- # Graph Transformer Steps
+::: cognite.neat.workflows.steps.lib.graph_transformer -->
