+++ tests/tests_unit/app/api/test_workflows/Extract RDF Graph and Generate Assets_workflow.yml
@@ -1,0 +1,3 @@
+assets: []
+labels: []
+relationships: []
