--- tests/tests_unit/app/api/test_workflows/extract_rdf_graph_generate_assets_workflow.yml
@@ -1,3 +1,0 @@
-assets: []
-labels: []
-relationships: []
